requirements.txt
pyserial
pygame
pymunk
mediapipe
opencv-python
pygetwindow